// front-end/src/app/models/generated/models/product.ts
/**
 * Products API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Photo } from './photo';


export interface Product { 
    id?: string | null;
    name?: string | null;
    description?: string | null;
    category?: string | null;
    price?: number | null;
    stock?: number;
    photos?: Array<Photo> | null;
    readonly imageUrl?: string | null;
    createdAt?: string;
    updatedAt?: string;
}