// front-end/src/app/models/generated/models/updateProductDto.ts
/**
 * Products API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateProductDto { 
    name?: string | null;
    description?: string | null;
    category?: string | null;
    price?: number | null;
    stock?: number | null;
    imageUrl?: string | null;
}